ANDROID: cpuset: Make cpusets restore on hotplug
This deliberately changes the behavior of the per-cpuset
cpus file to not be effected by hotplug. When a cpu is offlined,
it will be removed from the cpuset/cpus file. When a cpu is onlined,
if the cpuset originally requested that that cpu was part of the cpuset,
that cpu will be restored to the cpuset. The cpus files still
have to be hierachical, but the ranges no longer have to be out of
the currently online cpus, just the physically present cpus.

BUG: b:152026859
BUG: b:152062524
TEST: manually check cpuset through cpu onlining & offlining

(cherry picked from commit 95a20bd6ae349f770e573f73c69147cd2001e93b
 in v4.19 repo)

[AmitP: Refactored original changes to align with upstream commit
        201af4c0fab0 ("cgroup: move cgroup files under kernel/cgroup/")]
[jsbarnes: backport to 4.4, adjust for changed context in cpuset_init]
[sonnyrao: backport to 3.18, context changes]
Signed-off-by: Dmitry Shmidt <dimitrysh@google.com>
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Signed-off-by: Jesse Barnes <jsbarnes@google.com>
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Change-Id: Ia12a7ea4a2306710bc837ea91af823487724c014
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2133074