CHROMIUM: [media] mtk-mdp: Remove states for format checks
The mtk-mdp driver uses states to check if the formats have been set
on the capture and output when turning the streaming on, setting
controls or setting the selection rectangles.
Those states are reset when 0 buffers are requested like when checking
capabilities.
This patch removes all format checks and set one by default as queues in
V4L2 are expected to always have a format set.

https://linuxtv.org/downloads/v4l-dvb-apis/uapi/v4l/vidioc-streamon.html
https://linuxtv.org/downloads/v4l-dvb-apis/uapi/v4l/vidioc-g-ctrl.html
https://linuxtv.org/downloads/v4l-dvb-apis/uapi/v4l/vidioc-g-selection.html

BUG=chromium:1020303
TEST=Ran tast with video.DecodeAccel* on hana with crrev.com/c/1880208
TEST=Ran tast with video.DecodeAccel* on hana without crrev.com/c/1880208
TEST=Ran tast with video.Play.* on hana with crrev.com/c/1880208
TEST=Ran tast with video.Play.* on hana without crrev.com/c/1880208
TEST=Compiled and deployed latest Chrome on hana. Confirmed it isn't
     crashlooping during boot.

Change-Id: I0b67ba4677d5202a1690825ef13b1709973ae616
Signed-off-by: Francois Buergisser <fbuergisser@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1895149
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>