CHROMIUM: sched: track timestamps for hung task debugging
Having a list of stack traces without any notion
of which came first, isn't particularly useful.  It would
be nice to have some indication of an order of events.

We store the timestamps whenever a task goes to sleep or
it becomes runnable.  This should make it easier to construct
timelines from hung task crashes.

TEST=Use the provoke_crash/DIRECT interface in the debugfs to trigger
hung task crash.  Verified that the timestamps look sane.

BUG=chromium:246445

Change-Id: Idddec98d7d6f855f5065145ed2513ef2b65b45fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1992480
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Salman Qazi <sqazi@google.com>
Commit-Queue: Salman Qazi <sqazi@google.com>