BACKPORT: FROMLIST: bridge: multicast to unicast
Implements an optional, per bridge port flag and feature to deliver
multicast packets to any host on the according port via unicast
individually. This is done by copying the packet per host and
changing the multicast destination MAC to a unicast one accordingly.

multicast-to-unicast works on top of the multicast snooping feature of
the bridge. Which means unicast copies are only delivered to hosts which
are interested in it and signalized this via IGMP/MLD reports
previously.

This feature is intended for interface types which have a more reliable
and/or efficient way to deliver unicast packets than broadcast ones
(e.g. wifi).

However, it should only be enabled on interfaces where no IGMPv2/MLDv1
report suppression takes place. This feature is disabled by default.

The initial patch and idea is from Felix Fietkau.

Cc: Felix Fietkau <nbd@nbd.name>
Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>

This is a backport of following version of the patch
(am from https://patchwork.kernel.org/patch/9494055/)
commit 6db6f0eae6052b70885562e1733896647ec1d807
taken from this mailing list
https://github.com/openwrt-mirror/openwrt/commit/54b61868fca0d520fa6aec7cf80b6e5180b0849a

BUG=b:153689749
TEST=Build on AP and test with 4 different clients subscribing
to different mcast streams.

Signed-off-by: Kevin Hayes <kevinhayes@chromium.org>
Signed-off-by: Kishan Kunduru <kkunduru@chromium.org>
Change-Id: I390ff0517d9e6cc1e4e7be4fcf3a6fd3fe832faf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2150925
Reviewed-by: Hugo Benichi <hugobenichi@google.com>